% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_models.R
\name{print.vast_model_diff}
\alias{print.vast_model_diff}
\title{Print a model comparison}
\usage{
\method{print}{vast_model_diff}(x, ...)
}
\arguments{
\item{x}{A \code{vast_model_diff} object.}

\item{...}{Ignored.}
}
\description{
Print a model comparison
}
