% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.vast_translation}
\alias{print.vast_translation}
\title{Print method for vast_translation objects}
\usage{
\method{print}{vast_translation}(x, ...)
}
\arguments{
\item{x}{A \code{vast_translation} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for vast_translation objects
}
